Difficulties
So I ended up changing a bunch of stuff.  I want to end up with a header element that has another header element on the left side and a nav element on the right side.  The reason I'm doing this all within a header element is so I can easily move or remove the header from the flow of the page while maintaining the rest of the layout.  What I ended up doing was making a positioned <header> element and placing another <header> and a <nav> in that parent header element using absolute positioning.  It seems to work well, but the absolute positioning may become a problem later.  We'll see!