Working on the skeleton.
The absolute position trick seems to give me the effect I desire.  Right now I have the main skeleton of the page set up.  There is a header element with a left sided header element and a right sided nave element.  Then there is a main element with a figure, a left sided aside element and a right sided aside element.  Finally, there is a footer.