The issue of background image opacity
So, I spent a little while trying to solve an issue.  I want a partial opacity main content image with a paragraph/caption super imposed over it.  I tried using backgroundimage, but it was giving me trouble.  What I ended up doing was defining a figure div with the same height and width as an image element.   I then used the absolute positioning tag to absolutely position the paragraph within a few pixels of its container's top left corner.  Luckily, the container is the size of the image, so the paragraph really gets position a few pixels from the image's top left corner.