Testing z-index ***
I'm trying to understand the paragrah and img tags better.  I don't really understand why the paragraph is showing up on top of the img.  I tried messing with the z-index to see if I could get the image to lay on top of the paragraph tag, but the element is not floated and so z-index does not appear to be applying.  Interesting stuff.