Styling the page
Putting in a background and trying to get it so that I can get make an opaque image without the background's color bleeding into it.  What I will try to do is make a white div where I want the figure, and then I'll post the figure within that.  I think it will pick up the white from the div and the opacity will blend into that white color instead of the orange color of the background.  Going to make a separate branch to test this stuff out.