It works!
Nice, so it appears that if you wrap an image with a div with a white background, when you opaque it it will opaque to the white not the colored background as expected.  I actually used a whole new div to do it, but then I realized that the image is already wrapped in a figure element, so all I had to do was change the background-color of the figure element to white.   Could cause some specificity issues in the future, but I don't intend to have any other figures, so I think I'll be good for this project.