Got it
So the problem was as follows.  The body element is actually a child element to the html element.  Therefore, you need to tell the html element to take up 100% of the page so that the body will take up 100% of the page as well.  Then if you place stuff absolutely at thebottom of the body container it will be placed at the bottom of the page.  Luckily enough, this also solved a background image issue I was having.

One last take away from this experiment, it is better to use percentages on background-size instead of cover.  The cover definition is actually a little different then what you'd expect it to be.