Figured out how clearfix works.
So I was having a little bit of trouble getting the article div to appropriately wrap around the floated elements.  I used the clearfix, and I even think I understand what it is it's doing.

However, from what I'm reading, it seems like flexbox is the preferred method to style pages.  I think it'll probably be better to make the page with my float style in order to learn the old methods before I implement the new ones.